+++ src/main/scala/io/viash/platforms/nextflow/NextflowAuto.scala
@@ -32,11 +32,10 @@ case class NextflowAuto(
   @description(
     """If `true`, an output tuple containing a map with a File (e.g. `["foo", [ output: file("out.h5ad") ] ]`) is automatically transformed to a map (i.e. `["foo", file("out.h5ad")]`).
       |
-      |Default: `true`.
+      |Default: `false`.
       |""".stripMargin)
-  @default("True")
-  simplifyOutput: Boolean = true,
-  // TODO: change the default to false
+  @default("False")
+  simplifyOutput: Boolean = false,
 
   @description(
     """If `true`, the module's transcripts from `work/` are automatically published to `params.transcriptDir`.